requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langchain-community==0.0.10
requests==2.31.0
python-dotenv==1.0.0
pydantic>=1.10,<2.0
requests
joblib
scikit-learn
numpy